fix overriding of deprecated nonvariational methods
during refactorings, many lifted methods kept an unlifted counterpart as deprecated to
allow easy migration of nonvariational code. This was dangerous when other classes
overwrote the old methods, but not the new variational ones.
As a new policy, a new VDeprecated annotation was introduced and those methods
are marked final, forcing to migrate all overridden versions of those methods
to be lifted as well.