redesign of V.*map and V.*flatMap
V.vmap and V.vflatMap were not well defined. This propagated a context and could lead to calling the function with invalid contexts.

The API splits this now into two separate use cases that were previously not distinguished:

V.smap applies the map function to a restricted configuration space and returns only that restricted configuration space.
V.pmap applies the map function to a partial configuration space and a separate function (usually the identity function) to the remaining configuration space.

The former is just a shorthand for x.select(c).map(f), whereas the latter is a shorthand for V.choice(c, x.select(c).map(f), x.select(c.not).map(f')).
Both have different use cases. The former applies an operation in a context, whereas the latter updates part of a larger context.

An important previous usecase of vmap was to propagate a context. The new smap and pmap functions are overloaded that
they both work with functions that get or ignore the context.